requirements.txt
# Core ADK and A2A dependencies
google-adk>=1.0.0
# google-a2a-python>=1.0.0  # Not available on PyPI

# AI/ML Platform - Multiple LLM providers
anthropic>=0.8.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
openai>=1.0.0

# Language and accessibility analysis
langdetect>=1.0.9
textstat>=0.7.3
spacy>=3.7.0
nltk>=3.8.1

# Web accessibility testing
# axe-core-python>=4.7.0  # Not available on PyPI
selenium>=4.15.0
beautifulsoup4>=4.12.0

# Audio/Voice processing
speechrecognition>=3.10.0
pyttsx3>=2.90
# pyaudio>=0.2.11  # Can be problematic on Windows

# UI and visualization
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0

# Testing and validation
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Accessibility Testing Tools
playwright>=1.40.0
requests>=2.31.0
lxml>=4.9.0

# Alternative accessibility tools
axe-selenium-python>=2.1.6
# accessibility-checker>=3.1.0  # Not available on PyPI

# Performance monitoring
prometheus-client>=0.19.0
psutil>=5.9.0

# Configuration and utilities
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Development tools
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Additional useful packages
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
colorama>=0.4.6
tqdm>=4.65.0 